Removing remaining unused recording parts
