+++ src-tauri/Cargo.toml
@@ -40,7 +40,6 @@ futures = "0.3.31"
 xcap = "0.6.0"
 uuid = { version = "1.17.0", features = ["v4"] }
 chrono = "0.4.41"
-hound = "3.5.1"
 ffmpeg-sidecar = "2.0.6"
 rmp-serde = "1.3.0"
 tauri-plugin-os = "2"
