+++ src-tauri/src/recording/mod.rs
@@ -1,8 +1,6 @@
 mod audio;
-mod camera;
 pub mod commands;
-mod ffmpeg;
 mod file;
 mod input_events;
 pub mod models;
-mod screen;
+mod video;
